rewrite the bsmdoc_listbullet function.
For nested listbullet,e.g.,
* item
** child

Before, the generate html code would look like
<ul>
<li>item</li>
<ul>
<li>item</li>
</ul>
</ul>
now it will look like
<ul>
<li>item
<ul>
<li>item</li>
</ul>
</li>
</ul>

Also add "<p>" tag to paragraph not followed by "new paragraph" symbol
('\n\n'). It may cause conflicts. For example, the text in the function block
may not be a paragraph. Thus, we only add the "tag" to the text not in a
function block and followed by '\n'.