Support generic style for html tags.
For example, you can add 'id' to a div function block
{!div|id="mydiv"||
!}
For each function block argument, if it does not have any '=', it will be
treated as a 'class'; otherwise, it will keep untouched and be added to the
tag.

Add video function block.

Add side-menu to index.html. And the main content of index.bsmdoc is moved to
content.bsmdoc for clarity.